+++ enlighten/Controller.py
@@ -938,6 +938,7 @@ class Controller:
                          self.laser_control,
                          self.laser_temperature,
                          self.laser_watchdog,
+                         self.plugin_controller,
                          self.raman_shift_correction,
                          self.raman_intensity_correction,
                          self.reference_feature,
@@ -1804,7 +1805,7 @@ class Controller:
         # lesser value and reserved 0xffff for "frame markers". Raw should match 
         # processed at this point, but "raw" is preferred to clarify saturation 
         # is irrespective of dark correction or any post-processing.
-        if pr.raw.max() >= 0xfffe: 
+        if pr.raw.max() >= 0xfffe and not spec.settings.is_ids():
             # todo: self.status_indicators.detector_warning("detector saturated")
             self.marquee.error("detector saturated")