+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Version History
 
+- 2025-04-04 4.1.25
+    - IDS
+        - added IDSPeak plugin
+        - update AreaScan to support .png files
+    - Area Scan
+        - deprecated 'Fast' checkbox
+        - added 'Normalize' checkbox
+        - trying to add start/stop lines to image
+    - plugins
+        - added update_visibility method to handle spectrometer hotswaps etc
+        - fixed Prod.EmissionLines unit
 - 2025-02-28 4.1.24
     - plugins
         - automatically add "Copy to Clipboard" button for plugins generating a dataframe